requirements.txt
pandas
numpy
librosa
colorthief
pillow